I3 autorun script for power status
script nofity-bat.sh uses notify-send and my two freaking
batteries to do some basic power notification.
If the power drops below a certain level, 10%, it makes a critical
message with the current battery level. If the power is inbetween
two thresholds, 10-15%, a normal message is raised.

Currently the existing power is the simple average of the two
percentages. May look at combining the total energy from each
battery and the current energy level. But that will be for a later
day if I'm getting real pedantic

This script is called at start by i3