POLYBAR Use polybat.sh for battery status
New script can be used to display a single battery percentage
in the polybar. Takes the average of the two batteries and
displays that plus and icon.
Icon is a lightning bolt if charging, skull if under 15%,
and nothing otherwise.

Polybar is configured to run this script every 90 seconds. The delay
is a bit long to properly register events. Should maybe look into
having this be a perpetual script instead.

Maybe merge notify-bat.sh and polybat.sh into one script?
Have polybar call a script that also notifies?