BASH Cleaner shell prompt - little git integration
Cleaned up the shell prompt to a very minmal form.
Very close to the default, but with some better git stuff.
If you are in a git repository, the branch will be added to the
prompt. If the repository is "clean", e.g. no staged changes nor
modifications, then a green check is added. Otherwise, two
additional symbols are used to display the precense of staged
changes and unstaged changes. These are yellow and red, respectively